requirements.txt
# EventAI — Python dependencies
# Install: pip install -r requirements.txt
#
# face_recognition requires cmake + dlib.
# On Ubuntu/Debian: sudo apt install cmake build-essential libopenblas-dev
# On macOS (Homebrew): brew install cmake

fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9       # required for Form / UploadFile
motor==3.4.0                  # async MongoDB driver
pymongo==4.7.2                # used by motor internally
python-dotenv==1.0.1
authlib==1.3.1
httpx==0.27.0                 # authlib transport
itsdangerous==2.2.0           # SessionMiddleware
requests==2.32.3
Pillow==10.3.0
numpy==1.26.4
face-recognition==1.3.0       # dlib CNN backend
